update upload field
