+++ src/constants/marketplace.ts
@@ -23,31 +23,3 @@ export const defaultMetadataDetail: MetadataDetail = {
     from: 'designer',
     id: -1,
 };
-
-export interface Metadata {
-    name: Translation;
-    desc: Translation;
-    contributor: string;
-    lastUpdateOn: string;
-    type: 'MiscNode' | 'Station';
-    status: 'public' | 'pending' | 'rejected';
-    paramStr: string;
-    svgString: string;
-    svgHash: string;
-}
-
-export interface Marketplace {
-    [id: number]: Metadata;
-}
-
-export interface ResponseMetadata {
-    id: number;
-    data: string;
-    hash: string;
-    lastUpdateAt: string;
-    status: 'public' | 'pending' | 'rejected';
-    type: 'MiscNode' | 'Station';
-    userId: number;
-    svgString: string;
-    svgHash: string;
-}
